Models/Chat_model/Antropic.py
from langchain_anthropic import ChatAnthropic
from dotenv import load_dotenv

load_dotenv()

model = ChatAnthropic(model='claude-3-5-sonnet-20241022')

query=input("Ask AI:")
result=model.invoke(query)
print(result.content)